idemo pops up on screen and we have carbon event loop going. at present, though, nothing works. carbevent.cpp is a temporary version for experimenting while avoiding many #if carbon
git-svn-id: http://www.neuron.yale.edu/svn/neuron/iv/trunk@1365 a16e031d-7b14-0410-9afa-d909082ea83f

+++ src/lib/IV-Mac/font.cpp
@@ -581,7 +581,7 @@ int Font::index(
 	int xoffset = Session::instance()->default_display()->to_pixels(Coord(offset));
 	w = 0;
 	for (p = s, n = 0; *p != '\0' && n < len; ++p, ++n) {
-		cw = width(*p);
+		cw = (int)width(*p);
 		w += cw;
 		if (w > xoffset) {
 			break;
